adapt server to the new vector API
